; Conflicts:
;	src/hexdoc/__version__.py
;	src/hexdoc/_templates/pages/patchouli/multiblock.html.jinja
;	src/hexdoc/cli/ci.py
;	src/hexdoc/patchouli/page/pages.py

+++ src/hexdoc/cli/ci.py
@@ -29,6 +29,7 @@ def build(
     props_file: PropsOption,
     release: ReleaseOption,
     clean_exports: bool = True,
+    run_hatch_build: bool = True,
 ):
     from . import app as hexdoc_app
 
@@ -56,12 +57,13 @@ def build(
         clean_exports=clean_exports,
     )
 
-    site_dist = site_path / "dist"
-    if site_dist.is_dir():
-        shutil.rmtree(site_dist)
+    if run_hatch_build:
+        site_dist = site_path / "dist"
+        if site_dist.is_dir():
+            shutil.rmtree(site_dist)
 
-    subprocess.run(["hatch", "build", "--clean"], check=True)
-    shutil.copytree("dist", site_dist)
+        subprocess.run(["hatch", "build", "--clean"], check=True)
+        shutil.copytree("dist", site_dist)
 
     env.set_output("pages-url", pages_url)
